requirements.txt
flask
numpy
pandas
mysql-connector-python
Flask-Session